A prototype of the core method of the library (`#take_snapshot`)
`#take_snapshot` is the core method of the library.

`SnapshotUI::Test::RackTestHelpers` is intended to be included in a test file of a rack application where a user wishes to take snapshots of UI responses.

In the first implementation of `#test_snapshot` we simply print out the body of the provided response.