Display a single taken snapshot
We introduce a `/response/:slug` and `/response/raw/:slug` routes.

The `/response/raw/:slug` route renders a raw html of a response body.

The `/response/:slug` route renders an iframe with the raw response. Making a distinction between sets the stage for later when we will do live reloads when the taken response body changes.