Replace the command `snapshot_ui watch` with `snapshot_ui live`
This commit introduces a breaking change, but since there is only one user of the library it's acceptable.

The process responsible for broadcasting live refresh updates now has access to the Snapshot UI configuration.

This resolves some code duplication and the possibility of misconfiguration that existed previously.

Since we cannot know where the Snapshot UI configuration file will be located, the user is required to pass the path to it as an argument.

As a default port for a websocket URL we use 49152 since it is quite unlikely it will already be occupied.